internal/apiserver/route_delete_contract_subscription.go
// Copyright © 2021 Kaleido, Inc.
//
// SPDX-License-Identifier: Apache-2.0
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package apiserver

import (
	"net/http"

	"github.com/hyperledger/firefly/internal/config"
	"github.com/hyperledger/firefly/internal/i18n"
	"github.com/hyperledger/firefly/internal/oapispec"
	"github.com/hyperledger/firefly/pkg/fftypes"
)

var deleteContractSubscription = &oapispec.Route{
	Name:   "deleteContractSubscription",
	Path:   "namespaces/{ns}/contracts/subscriptions/{subid}",
	Method: http.MethodDelete,
	PathParams: []*oapispec.PathParam{
		{Name: "ns", ExampleFromConf: config.NamespacesDefault, Description: i18n.MsgTBD},
		{Name: "subid", Description: i18n.MsgTBD},
	},
	QueryParams:     nil,
	FilterFactory:   nil,
	Description:     i18n.MsgTBD,
	JSONInputValue:  nil,
	JSONInputMask:   nil,
	JSONOutputValue: nil,
	JSONOutputCodes: []int{http.StatusNoContent}, // Sync operation, no output
	JSONHandler: func(r *oapispec.APIRequest) (output interface{}, err error) {
		subid, err := fftypes.ParseUUID(r.Ctx, r.PP["subid"])
		if err != nil {
			return nil, err
		}
		err = r.Or.Contracts().DeleteContractSubscriptionByID(r.Ctx, subid)
		return nil, err
	},
}
